% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortTables.R
\name{createCohorts}
\alias{createCohorts}
\title{Create cohort tables}
\usage{
createCohorts(
  connectionDetails,
  cohortIds,
  baseUrl,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  jsonFolder = file.path("inst/cohorts"),
  sqlFolder = file.path("inst/sql/sql_server"),
  getCohortCounts = FALSE
)
}
\arguments{
\item{connectionDetails}{Connection details to the database server using createConnectionDetails() from the DatabaseConnector package}

\item{cohortIds}{List of cohort ids to be included in the cohort table}

\item{baseUrl}{The base URL for the Atlas Web Api. For example, "http://api.ohdsi.org:8080/WebAPI"}

\item{cdmDatabaseSchema}{Schema name for the OMOP CDM database}

\item{cohortDatabaseSchema}{Schema name for the cohort table}

\item{cohortTable}{Name for the cohort table, default is "cohort"}

\item{jsonFolder}{JSON folder location for the cohort definition set}

\item{sqlFolder}{SQL folder location for the cohort definition set}

\item{getCohortCounts}{When TRUE, it will return cohort counts from the cohort table just created}
}
\description{
Generate cohort tables using the Atlas Web Api cohort id(s) and saving the cohort definition set to the to a JSON or SQL folder
}
\examples{
\dontrun{
createCohorts(
    connectionDetails = Eunomia::getEunomiaConnectionDetails(),
    cohortIds = cohortIds,
    baseUrl = "http://api.ohdsi.org:8080/WebAPI",
    cdmDatabaseSchema = "main",
    cohortDatabaseSchema = "main",
    cohortTable = "example_cohort"
)
}

}
