% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortTables.R
\name{dropCohorts}
\alias{dropCohorts}
\title{Cleaning up and Dropping Cohort Tables}
\usage{
dropCohorts(connectionDetails, cohortDatabaseSchema, cohortTable = "cohort")
}
\arguments{
\item{connectionDetails}{Connection details to the database server using createConnectionDetails() from the DatabaseConnector package}

\item{cohortDatabaseSchema}{Schema name for the cohort table}

\item{cohortTable}{Name for the cohort table, default is "cohort"}
}
\description{
Removing the cohort and residual tables
}
