% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasUtilities.R
\name{get_concept}
\alias{get_concept}
\title{Takes a valid OMOP Concept ID and returns a list of concept fields}
\usage{
get_concept(id)
}
\arguments{
\item{id}{Integer representing an OMOP Concept ID}
}
\value{
list Fields related to the concept
}
\description{
Takes a valid OMOP Concept ID and returns a list of concept fields
}
\examples{
get_concept(436665)

}
