% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasUtilities.R
\name{get_concept_cohort}
\alias{get_concept_cohort}
\title{Getting all the relevant concept IDs related to each condition.
Takes a json filename corresponding to the condition and returns the condition concept set}
\usage{
get_concept_cohort(filename)
}
\arguments{
\item{filename}{JSON filename corresponding to the condition
(i.e bipolar disorder, depression, suicidality)}
}
\value{
DataFrame with descriptions of the related concept IDs
}
\description{
Getting all the relevant concept IDs related to each condition.
Takes a json filename corresponding to the condition and returns the condition concept set
}
\examples{
## Retrieving the dataframe of related concepts associated with bipolar disorder
filename <- "../data/exp_raw/concept_sets/bipolar_concept.json"
get_concept_cohort(filename)

# | CONCEPT_ID | CONCEPT_NAME         | ... | VOCABULARY_ID    | CONCEPT_CLASS_ID  |
# | ---------- | -------------------- | --- | ---------------- | ----------------- |
# | 432876     | Bipolar I disorder   | ... | SNOMED           | Clinical Finding  |
# | 3220652    | Bipolar 1 disorder   | ... | Nebraska Lexicon | Clinical Finding  |

}
